feat: add second redeposit user
