+++ helpers/utils.py
@@ -1,0 +1,7 @@
+def approx(actual, expected, percentage_threshold):
+    print(actual, expected, percentage_threshold)
+    diff = int(abs(actual - expected))
+    # 0 diff should automtically be a match
+    if diff == 0:
+        return True
+    return diff < (actual * percentage_threshold // 100)
